Create new client to get register data
This commit allows clients to obtain a register's records (current, expired, or both), obtain a register's history and obtain the metadata for the register (including which fields comprise it and the custodian responsible for maintaining it). Each register is separately managed by a new `RegisterClient`, and may be updated manually with a call to its `refresh_data` method.