Cache register data
This commit adds mini_cache to the client library and uses it to cache results from the call to `/download-rsf` for a given `RegisterClient` object. The cache duration is configurable by the caller when creating a new `RegistersClient`; if no duration is specified by the caller, a default duration of one minute is assumed. All `RegisterClient`s will use the provided cache duration.